Add a error dialog, if a ui cannot commit after x attemps (#151)
