+++ web/src/app/modules/notification/modules/modals/services/confirmation-modal.service.ts
@@ -1,17 +1,20 @@
 import { Injectable } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
+import { ValidationService } from 'src/app/modules/forms/modules/validation/services/validation.service';
 import { SpecmateDataService } from '../../../../data/modules/data-service/services/specmate-data.service';
 import { TypedModalContent } from '../components/typed-modal-content.component';
 import { Dialogtype } from '../modal-dialog-type';
-import { ValidationService } from 'src/app/modules/forms/modules/validation/services/validation.service';
 import { ModalService } from './modal-service';
+import { SimpleModal } from './simple-modal.service';
 
 @Injectable()
-export class ConfirmationModal {
-    constructor(private modalService: ModalService,
+export class ConfirmationModal extends SimpleModal {
+    constructor(protected modalService: ModalService,
         private dataService: SpecmateDataService,
         private translate: TranslateService,
-        private validator: ValidationService) { }
+        private validator: ValidationService) {
+        super(modalService);
+    }
 
     public async openSave(message: string, withCancel = true, withDiscard = true): Promise<any> {
         const modalRef = this.modalService.open(TypedModalContent);
@@ -26,18 +29,6 @@ export class ConfirmationModal {
     }
 
     public confirmDelete(title: string, message: string): Promise<any> {
-        const modalRef = this.modalService.open(TypedModalContent);
-        modalRef.componentInstance.options = Dialogtype.okCancelDialog(title, message);
-        return modalRef.result;
-    }
-
-    public openOk(title: string, message: string): Promise<any> {
-        const modalRef = this.modalService.open(TypedModalContent);
-        modalRef.componentInstance.options = Dialogtype.okDialog(title, message);
-        return modalRef.result;
-    }
-
-    public openOkCancel(title: string, message: string): Promise<any> {
         const modalRef = this.modalService.open(TypedModalContent);
         modalRef.componentInstance.options = Dialogtype.okCancelDialog(title, message);
         return modalRef.result;
