+++ web/src/assets/i18n/de.json
@@ -215,6 +215,7 @@
 		"continue": "Fortfahren nicht möglich, da die folgende Fehler existieren:",
 		"discard": "Das Speichern ist nicht möglich, da die folgenden Fehler existieren. Änderungen verwerfen?",
 		"message": "Das Speichern ist nicht möglich, da die folgenden Fehler existieren:",
+		"retry": "Fehler beim Speichern. Bitte versuchen Sie es später erneut.",
 		"title": "Speichern nicht möglich"
 	},
 	"search": "suchen",
